requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
babel==2.9.0
    # via
    #   django-oscar
    #   django-phonenumber-field
certifi==2020.12.5
    # via requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.2.1
    # via
    #   -r requirements/base.in
    #   social-auth-core
defusedxml==0.6.0
    # via
    #   python3-openid
    #   social-auth-core
django-appconf==1.0.4
    # via django-compressor
django-compressor==2.4
    # via -r requirements/base.in
django-crum==0.7.9
    # via -r requirements/base.in
django-extra-views==0.11.0
    # via django-oscar
django-haystack==2.8.1
    # via django-oscar
django-oscar==2.0.4
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
django-phonenumber-field==2.0.1
    # via django-oscar
django-tables2==1.21.2
    # via django-oscar
django-treebeard==4.3.1
    # via django-oscar
django-waffle==2.0.0
    # via -r requirements/base.in
django-widget-tweaks==1.4.8
    # via django-oscar
django==2.2.17
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   django-appconf
    #   django-crum
    #   django-extra-views
    #   django-haystack
    #   django-oscar
    #   django-phonenumber-field
    #   django-tables2
    #   django-treebeard
    #   djangorestframework
    #   edx-auth-backends
    #   jsonfield2
djangorestframework==3.11.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
edx-auth-backends==3.0.2
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
factory-boy==2.12.0
    # via django-oscar
faker==5.0.0
    # via factory-boy
idna==2.10
    # via requests
jsonfield2==3.0.3
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
oauthlib==3.1.0
    # via
    #   requests-oauthlib
    #   social-auth-core
phonenumbers==8.12.16
    # via
    #   django-oscar
    #   django-phonenumber-field
pillow==7.2.0
    # via django-oscar
purl==1.5
    # via django-oscar
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via
    #   edx-auth-backends
    #   social-auth-core
python-dateutil==2.8.1
    # via faker
python3-openid==3.2.0
    # via social-auth-core
pytz==2020.5
    # via
    #   babel
    #   django
qrcode==6.1
    # via -r requirements/base.in
rcssmin==1.0.6
    # via django-compressor
requests-oauthlib==1.3.0
    # via social-auth-core
requests==2.25.1
    # via
    #   requests-oauthlib
    #   social-auth-core
rjsmin==1.1.0
    # via django-compressor
six==1.15.0
    # via
    #   cryptography
    #   django-compressor
    #   django-extra-views
    #   edx-auth-backends
    #   purl
    #   python-dateutil
    #   qrcode
    #   social-auth-app-django
    #   social-auth-core
social-auth-app-django==3.4.0
    # via edx-auth-backends
social-auth-core==3.4.0
    # via
    #   edx-auth-backends
    #   social-auth-app-django
sorl-thumbnail==12.7.0
    # via -r requirements/base.in
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
urllib3==1.26.2
    # via requests
